Using Makers to trace a custom path
Basic implementation of an 2by1 array of visual markers for path planning. Markers contain own poses and publish their poses on a /visual_marker topic.  The idea is to set robot´s goals equal to the marker´s pose, hence getting the error different of the path.

EDIT: For the path planner the teb_local_planner is being analysed